Fixed bug "startup.lua:82: attempt to index field '?' (a nil value)"
Earlier the code was acting on commands which were not defined in it. So
whenever someone used something like !getprofiles the script still ran
and failed at line 82.
Also changed condition for no arguments from sData:len() == 0 to not
tBreak[1]  to provide more robustness ( for ex if someone sends !check
with lots of space at the end we will get error with the former
condition)